Refactor before adding tests
* Pull WMATA API handling out into its own module for a better interface
* Expose created MentroTransit instance to avoid potentially creating multiple instances.
