+++ lib/metro-transit.js
@@ -1,31 +1,14 @@
 var _ = require('lodash');
-var request = require('request');
 
 var props = require('../resources/properties');
 var stations = require('../resources/stations');
-
 var secrets = require('../secret/secrets');
 
 var AlexaSkill = require('./AlexaSkill');
 var utils = require('./utils');
+var WmataApi = require('./api-wmata');
 
 var APP_ID = secrets.applicationId;
-
-var wmataReq = request.defaults({
-  baseUrl: props.wmata_base_url,
-  qs: secrets.apiKey
-});
-
-var getWmataResponse = function(endpoint, response, callback) {
-  wmataReq(endpoint, function(error, res, body) {
-    if (!error && res.statusCode === 200) {
-      callback(JSON.parse(body));
-    } else {
-      console.error(endpoint.concat(': Error with WMATA'), error);
-      response.tell(props.wmataErrorSpeechOutput);
-    }
-  });
-};
 
 var MetroTransit = function () {
   AlexaSkill.call(this, APP_ID);
@@ -64,7 +47,7 @@ MetroTransit.prototype.intentHandlers = {
     if (_.has(stations, stationName)) {
       var stationCode = stations[stationName].Code;
 
-      getWmataResponse(props.stationArrivalWmataUrl + stationCode, response, function(body) {
+      WmataApi.getWmataResponse(props.stationArrivalWmataUrl + stationCode, response, function(body) {
         var trainArrivals = _.reduce(body.Trains, function(result, train) {
           if (train.DestinationName === 'Train' || train.DestinationName == 'No Passenger') return result;
 
@@ -104,7 +87,7 @@ MetroTransit.prototype.intentHandlers = {
   },
 
   GetServiceAdvisories: function(intent, session, response) {
-    getWmataResponse(props.serviceAdvisoriesWmataUrl, response, function(body) {
+    WmataApi.getWmataResponse(props.serviceAdvisoriesWmataUrl, response, function(body) {
       var incidents = body.Incidents;
       if (!incidents) {
         console.error(props.serviceAdvisoriesErrorConsole);
@@ -119,6 +102,4 @@ MetroTransit.prototype.intentHandlers = {
   }
 };
 
-module.exports = {
-  MetroTransit: MetroTransit
-};
+module.exports = new MetroTransit();
